# docs/swagger.yaml
basePath: /
definitions:
  controller.MessageRequestUnion:
    properties:
      direct:
        $ref: '#/definitions/dto.DirectMessageRequest'
      team:
        $ref: '#/definitions/dto.TeamMessageRequest'
    type: object
...
  dto.DirectMessageRequest:
    properties:
      receiverId:
        type: string
      senderId:
        type: string
      textContent:
        type: string
    type: object
  dto.FriendRequestListResponse:
    properties:
      requests:
        items:
          $ref: '#/definitions/dto.FriendRequestResponse'
        type: array
    type: object
  dto.FriendRequestResponse:
    properties:
      createdAt:
        type: string
      fromUserId:
        type: string
      status:
        type: string
      toUserId:
        type: string
    type: object
...
  dto.MessageDTO:
    properties:
      id:
        type: string
      receiverId:
        type: string
      senderId:
        type: string
      sentAt:
        type: string
      teamId:
        type: string
      textContent:
        type: string
    type: object
...
  dto.RespondFriendRequestRequest:
    properties:
      accept:
        type: boolean
    type: object
...
  dto.TeamMessageRequest:
    properties:
      senderId:
        type: string
      teamId:
        type: string
      textContent:
        type: string
    type: object
...
info:
  contact: {}
  title: StudyWithMe API
  version: "1.0"
paths:
...
  /friend-requests/{fromUserId}/{toUserId}:
    post:
      description: Send a friend request from one user to another
      parameters:
      - description: Sender User ID
        in: path
        name: fromUserId
        required: true
        type: string
      - description: Recipient User ID
        in: path
        name: toUserId
        required: true
        type: string
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Send a friend request
      tags:
      - friend-requests
    put:
      description: Accept or deny a friend request
      parameters:
      - description: Sender User ID
        in: path
        name: fromUserId
        required: true
        type: string
      - description: Recipient User ID
        in: path
        name: toUserId
        required: true
        type: string
      - description: Accept or deny
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RespondFriendRequestRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Respond to a friend request
      tags:
      - friend-requests
  /friend-requests/{userId}:
    get:
      description: Get pending friend requests for a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FriendRequestListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get pending friend requests
      tags:
      - friend-requests
  /messages:
    get:
      consumes:
      - application/json
      description: Get messages between 2 users or within a team
      parameters:
      - description: Messages type (direct/team)
        in: query
        name: type
        required: true
        type: string
      - description: User1 ID (direct message)
        in: query
        name: user1Id
        type: string
      - description: User2 ID (direct message)
        in: query
        name: user2Id
        type: string
      - description: Team ID (team message)
        in: query
        name: teamId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.MessageDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get all messages
      tags:
      - messages
    post:
      consumes:
      - application/json
      description: Create and send a message either to another user or to a team
      parameters:
      - description: Message type (direct/team)
        in: query
        name: type
        required: true
        type: string
      - description: The message request (this is only for documentation purposes,
          the actual request should be either DirectMessageRequest or TeamMessageRequest)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.MessageRequestUnion'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MessageDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Create and send a message
      tags:
      - messages
  /messages/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: The message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageDTO'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get a message by ID
      tags:
      - messages
  /messages/connect:
    get:
      responses:
        "101":
          description: Switching Protocols - WebSocket connection established
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Connect the user to the message WebSocket
      tags:
      - messages
...
securityDefinitions:
  Bearer:
    description: 'Bearer JWT token. Example: "Bearer YOUR_TOKEN"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"